core/src/test/java/com/avolution/actor/SupervisorTest.java
package com.avolution.actor;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.logging.Logger;

import static org.junit.jupiter.api.Assertions.*;

class SupervisorTest {

    private Supervisor supervisor;
    private AbstractActor actor;

    @BeforeEach
    void setUp() {
        supervisor = new Supervisor(Supervisor.SupervisionStrategy.RESTART);
        actor = new BasicActor();
        supervisor.addActor(actor);
    }

    @Test
    void testAddActor() {
        AbstractActor newActor = new BasicActor();
        supervisor.addActor(newActor);
        // Add assertions to verify actor addition
    }

    @Test
    void testHandleFailure() {
        Throwable cause = new RuntimeException("Test error");
        supervisor.handleFailure(actor, cause);
        // Add assertions to verify failure handling
    }

    @Test
    void testRestartActor() {
        supervisor.handleFailure(actor, new RuntimeException("Test error"));
        // Add assertions to verify actor restart
    }

    @Test
    void testStopActor() {
        supervisor = new Supervisor(Supervisor.SupervisionStrategy.STOP);
        supervisor.handleFailure(actor, new RuntimeException("Test error"));
        // Add assertions to verify actor stop
    }

    @Test
    void testResumeActor() {
        supervisor = new Supervisor(Supervisor.SupervisionStrategy.RESUME);
        supervisor.handleFailure(actor, new RuntimeException("Test error"));
        // Add assertions to verify actor resume
    }

    @Test
    void testHandleErrorPropagation() {
        AbstractActor child = new BasicActor();
        actor.addChild(child);
        Throwable cause = new RuntimeException("Test error");
        child.propagateError(cause);
        // Add assertions to verify error propagation
    }
}
